change all BigDecimals to Double in Uhrenverwaltung
replace bigDecimal with double

+++ src/main/java/de/watchmywatch/Uhrenverwaltung/Bracelet.java
@@ -1,14 +1,12 @@
 package de.watchmywatch.Uhrenverwaltung;
 
-import java.math.BigDecimal;
-
 public class Bracelet extends Watchpart
 {
     private double size;
     private ConnectionType connection;
 
     public Bracelet(Manufacturer manufacturer, String manufacturerPartID,
-                    Material material, int amountAvailable, BigDecimal price,
+                    Material material, int amountAvailable, double price,
                     double size, ConnectionType connection)
     {
         super(manufacturer, manufacturerPartID, material, amountAvailable, PartType.BRACELET, price);
